requirements.txt
beautifulsoup4
requests
db-sqlite3
flask
line-bot-sdkl
